GUI
A lot of work, but here we are.

To modify code, or to add code, go to `src\java\com\papa247\john\`. That is where the code is stored (other than the FXML resources for windows).

Oh and we're now a JavaFX application, so that means GUI.
I'll post a guide how to easily modify windows and what not later, for now check out "helpful links".

You'll need JDK15 for this project.
Please download and install JDK15.

Add an example data base that shows the basics steps in saving/loading JSON data.

Added the Google Java Style XML document so we can follow the style easily.