GUI: Address, Listing, Lease, Icons [1/2]
Part **1** of (hopefully) **2**.
This is the _addition_ of the Address, Listing, Lease, and the ListingsHost (and it's dependencies)

The UI on all objects, (minus Lease, which lacks a "browse" button), is setup. Only the logic is left to be completed!

Part two will focus on the logic (the controllers) of these GUI objects (windows/scenes).

**Changes:**

POM: Added 4 dependencies:
	org.kordamp.ikonli:
		ikonli-fontawesome5-pack
		ikonli-material-pack
		ikonli-javafx
		ikonli-core

App:
	Can now run in GUI mode or Text mode (change GUI to false)

ConsoleUI:
	The text based UI driver.

Enumerators *:
	Added icons to a few of them. (.toIcon())

Address:
	Fix up JSON conversion
	+ getRating()
	+ isEmpty()

Listing:
	Added setters (will do the same for address later)
	Fixed up JSON conversion

+ Address (GUI)
+ Listing (GUI) (theoretically done)
+ Lease (GUI)
+ SearchPane (GUI)
+ ListingHost (GUI) (theoretically done)
+ SetAmminities (GUI) (theoretically done)

User:
	Fixed .getRating() to point to the reviews of the user rather than the reviews written by the user...

Module-info:
	See pom.

Style:
	Icons!

UserWindow:
	Set default buttons (to login, register, apply)
	Fix inconstant button sizing (in register view)